requirements.txt
alembic==1.8.1
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
click==7.1.2
cryptography==37.0.4
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.1.3
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.2.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.30
mysql-connector-repackaged==0.3.1
protobuf==3.20.1
psycopg==3.1
psycopg2-binary==2.9.5
pycparser==2.21
PyMySQL==1.0.2
requests==2.24.0
SQLAlchemy==1.3.19
typing_extensions==4.3.0
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
